Replace codebase with swagger generated client
This replaces the code base completely, as the
new swagger based client is very different
from the current client code-wise.

This implements the base client as well as
a projects subclient with basic CRUD methods.
This is not finished at all at this stage but
builds the foundation for the structure.

Along come some integration tests using a
Makefile, scripts to setup and teardown
Harbor instances and a go test file showcasing
integration tests against that Harbor instance.