bump auto-generated models, fix project client
This PR bumps the client versions of both the `v1` and `v2` client to their current latest version (`v1.10.5` & `v2.1.0`)
Code-generation scripts have been adjusted to the above & auto-generated code was re-generated using the above versions.

Fixed a bug (+test case) when listing projects, that would return no error when the received slice is of length 0.