implicit storageLimit values
This PR extends the `quota` clients functionality in a way that invalid
provided values are implicitly set to an infinite storage limit (`-1`).
A value is considered invalid, if it is **smaller or equal to 0**.
The harbor API does not return errors when updating quotas to an invalid
value - it returns a `status 200` response without changing the quota value.